added fake data through generatefixtures.py Able to generate queries using fake data from graphql. code refactoring as per review. changes to running dev settings.
