+++ backend/hct_mis_api/apps/targeting/fixtures.py
@@ -1,35 +1,26 @@
 import datetime as dt
-import json
 
 import factory
 from account.fixtures import UserFactory
+from core.utils import JSONFactory
 from factory import fuzzy
 from household.fixtures import HouseholdFactory
 from targeting.models import TargetPopulation
 from targeting.models import TargetRule
 
 
-class JSONFactory(factory.DictFactory):
-    """A Json factory class to get JSON strings."""
-
-    @classmethod
-    def _generate(cls, create, attrs):
-        obj_dict = super()._generate(create, attrs)
-        return json.dumps(obj_dict)
-
-
 class TargetPopulationFactory(factory.DjangoModelFactory):
     class Meta:
         model = TargetPopulation
 
-    name = factory.LazyAttribute(lambda: "sentence")
+    name = "sentence"
     created_by = factory.SubFactory(UserFactory)
     last_edited_at = fuzzy.FuzzyDateTime(dt.datetime.utcnow().astimezone())
     status = factory.fuzzy.FuzzyChoice(
         TargetPopulation.STATE_CHOICES, getter=lambda x: x[0],
     )
-    total_households = factory.LazyAttribute(lambda: 10)
-    total_family_size = factory.LazyAttribute(lambda: 20)
+    total_households = 10
+    total_family_size = 20
 
     @factory.post_generation
     def households(self, create, extracted, **kwargs):
